Home ErrorEntry in error.ts, not the barrel
ErrorEntry wraps SerializedError and belongs beside it in error.ts. Moving it
there lets resource-types.ts import it from error.ts directly instead of from
the index barrel, so a foundational type module no longer depends on the
barrel. index.ts re-exports it, so package consumers are unchanged.
